--- client/src/features/auth/api/logoutUser.ts
@@ -1,6 +1,0 @@
-import { api } from '@/app/lib/api';
-
-export const logoutUser = async (): Promise<void> => {
-  const response = await api.post('/auth/logout');
-  return response.data;
-};
