+++ client/src/features/auth/api/useLoginMutation.ts
@@ -1,9 +1,10 @@
+import { api } from '@/app/lib/api';
 import { queryClient } from '@/app/lib/queryClient';
-import { getProfile } from '@/features/auth/api/getProfile';
-import { loginUser } from '@/features/auth/api/loginUser';
+import { getProfile } from '@/features/auth/api/useProfileQuery';
 import { useToast } from '@/shared/hooks/useToast';
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router';
+import { LoginFormData, LoginResponse } from '../types/login';
 
 export const useLoginMutation = () => {
   const navigate = useNavigate();
@@ -21,4 +22,9 @@ export const useLoginMutation = () => {
       showError('로그인에 실패했습니다. 다시 시도해주세요.');
     },
   });
+};
+
+const loginUser = async (loginData: LoginFormData): Promise<LoginResponse> => {
+  const response = await api.post('/auth/login', loginData);
+  return response.data;
 };
