--- client/src/features/auth/hooks/useCheckIfLoggedInQuery.ts
@@ -1,10 +1,0 @@
-import { checkIfLoggined } from '@/features/auth/api/checkIfLoggedIn';
-import { useQuery } from '@tanstack/react-query';
-
-export const useCheckIfLoggedInQuery = () => {
-  return useQuery({
-    queryKey: ['checkIfLoggedIn'],
-    queryFn: checkIfLoggined,
-    staleTime: 1000 * 60 * 10,
-  });
-};
