--- client/src/features/auth/hooks/useRandomNicknameQuery.ts
@@ -1,9 +1,0 @@
-import { getRandomNickname } from '@/features/auth/api/getRandomNickname';
-import { useQuery } from '@tanstack/react-query';
-
-export const useRandomNicknameQuery = () => {
-  return useQuery<string>({
-    queryKey: ['randomNickname'],
-    queryFn: getRandomNickname,
-  });
-};
